Created and drew a texture from a custom array of image data.
The struct Pixel holds 4 floats representing the rgba colour values of a pixel.

I then created an array that has as many elements as there are pixels in my defined screen size. It then defined a colour value for each pixel in the array.

I then passed the array to the glTexImage2D function which created a texture from the pixel data.

I can then draw that texture to the screen on top of a quad.