Significantly sped up the mandlebrot function.
I realised that if a pixel is surrounded by pixels in the mandlebrot set, then the pixel itself must also be within the set.

Therefore if you calculated every 2nd pixel on every row, while also having the starting points on the rows change from pos 0 to pos1 every odd row, then you would end up with an offset grid of calculated pixels for the set.

Now if you iterate over the remaining pixels that had yet to be calculated and checked if every pixel around it was already in the set, then you wouldn't need to do the calculation and can just immediately set it to be in the mandlebrot set.